requirements_local.txt
django==1.11
termcolor
unipath
fabric3
fabtools3
django-import-export==1.0.0
django-toolbelt==0.0.1
numpy==1.14
bokeh==0.12.2
